Change '.menu-anchor' to a button.
What do you think about changing the `.menu-anchor` from an anchor to a button?  Since it doesn't link to anywhere and doesn't function without JavaScript, it seems like it's better suited to be a `button` element. No JavaScript will need to be updated and as an added benefit the `button` element will be selected when tabbing.  Might want to update the class name if it's no longer an anchor.