syntax update and copyright year change
In the latest version of Compass, display(flex) has been changed to display-flex(flex) so updated in the footer sections - display(flex) is also used in a few other refills on this page but haven't personally used or tested them, might be broken there too. Also updated copyright year with javascript to keep it accurate in any year.