Start with a blank slate
https://github.com/thoughtbot/refills/projects/1#card-694321

We're rebuilding Refills with a narrower focus of components that would
specifically be helpful to the way designers at thoughtbot work.

This includes JavaScript and interaction-heavy components, but also a
focus on accessibility, unopinionated visual design, and lack of a grid
dependency.

Because we won't be keeping any components around as-is, and many of the
dependencies are outdated, it will be easier for us to start from a
clean-slate.