Support children in dropdown toggles
When clicking a child in a dropdown toggle, we were referring to that
child as the target, rather than referring to the button itself.

This change checks to see if the target is the toggle, and looks for a
parent toggle if it isn't.

Fixes https://github.com/thoughtbot/refills/issues/438